% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_factors.R
\name{demographic_factors}
\alias{demographic_factors}
\title{Demographic Factors}
\usage{
demographic_factors(
  local_authorities,
  variant_probabilities,
  start_year,
  end_year
)
}
\arguments{
\item{local_authorities}{a character vector listing the local authorities to select (use GSS local authority codes)}

\item{start_year}{the start year to use as the numerator for the factor (defaults to minimum year in data)}

\item{end_year}{the end year to use as the denomintar for the factor (defaults to the maximum year in data)}

\item{variant_probalities:}{the probability to select each of the variants, should be a named vector where the names
correspond to the possible variants}
}
\description{
Demographic Factors
}
\examples{
demographic_factors("E08000031", "principal")
demographic_factors("E08000031", "high migration")
}
