% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run Model}
\usage{
model(path, data, strategies, demog_factors_fn, strategy_params)
}
\arguments{
\item{path}{where to store the model results}

\item{data}{the data to run the model on, should contain four columns (rn, admiage, sex, speldur)}

\item{strategies}{a table containing two columns (rn, strategy). Multiple strategies may be defined per rn}

\item{demog_factors_fn}{a function which returns a table containing two columns (demogroup, factor)}

\item{strategy_params}{a table containing five columns:
(strategy, avoidance_low, avoidance_high, speldur_adjustment, strategy_weight)}
}
\value{
a table containing two columns (rn, speldur)
}
\description{
Run Model
}
