% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run Model}
\usage{
model(
  model_run_name,
  run_number,
  data,
  strategies,
  strategy_params,
  demog_factors_fn
)
}
\arguments{
\item{model_run_name}{what to name the files}

\item{data}{the data to run the model on, should contain four columns (rn, admiage, sex, speldur)}

\item{strategies}{a table containing two columns (rn, strategy). Multiple strategies may be defined per rn}

\item{strategy_params}{a table containing five columns:
(strategy, avoidance_low, avoidance_high, speldur_adjustment, strategy_weight)}

\item{demog_factors_fn}{a function which returns a table containing two columns (demogroup, factor)}
}
\value{
a table containing two columns (rn, speldur)
}
\description{
Run Model
}
